refactor(server): extract notification and bot logic to helper modules to resolve ESLint complexity and max-lines warnings
