+++ server/bot/services/chatbotServerService.js
@@ -110,6 +110,76 @@ function incrementServerRateCounter(userId) {
   }
 }
 
+// -- Helpers Internos --
+
+function _aggregateStock(stockData) {
+  const stockByMedicine = {}
+  for (const entry of (stockData || [])) {
+    stockByMedicine[entry.medicine_id] = (stockByMedicine[entry.medicine_id] || 0) + entry.quantity
+  }
+  return Object.entries(stockByMedicine).map(([medicine_id, quantity]) => ({
+    medicine_id,
+    quantity,
+  }))
+}
+
+function _filterTodayLogs(logsData, todayStr) {
+  return (logsData || []).filter(log => {
+    return getTodayLocal(getSaoPauloTime(log.taken_at)) === todayStr
+  })
+}
+
+async function _fetchRawPatientData(userId, yesterday) {
+  return Promise.all([
+    supabase
+      .from('medicines')
+      .select('id, name, dosage_per_pill, dosage_unit, active_ingredient, therapeutic_class')
+      .eq('user_id', userId),
+    supabase
+      .from('protocols')
+      .select('id, medicine_id, frequency, time_schedule, dosage_per_intake')
+      .eq('user_id', userId)
+      .eq('active', true),
+    supabase
+      .from('medicine_logs')
+      .select('protocol_id, taken_at')
+      .eq('user_id', userId)
+      .gte('taken_at', yesterday),
+    supabase
+      .from('stock')
+      .select('medicine_id, quantity')
+      .eq('user_id', userId)
+      .gt('quantity', 0),
+  ])
+}
+
+function _logQueryResults(userId, medicinesResult, protocolsResult, logsResult, stockResult) {
+  logger.debug('✅ Queries Supabase completadas', {
+    userId,
+    medicinesCount: medicinesResult.data?.length || 0,
+    medicinesError: medicinesResult.error?.message,
+    protocolsCount: protocolsResult.data?.length || 0,
+    protocolsError: protocolsResult.error?.message,
+    logsCount: logsResult.data?.length || 0,
+    logsError: logsResult.error?.message,
+    stockCount: stockResult.data?.length || 0,
+    stockError: stockResult.error?.message,
+  })
+}
+
+function _checkMedicineError(userId, medicinesResult) {
+  if (medicinesResult.error) {
+    const { code, details, hint } = medicinesResult.error
+    logger.error('❌ Erro ao buscar medicamentos', medicinesResult.error, {
+      userId,
+      errorCode: code,
+      errorDetails: details,
+      errorHint: hint,
+    })
+    throw medicinesResult.error
+  }
+}
+
 // -- Busca de dados do paciente --
 
 /**
@@ -132,69 +202,16 @@ export async function fetchPatientData(userId) {
 
   logger.debug('🔄 Executando 4 queries em paralelo (Supabase)', { userId })
 
-  const [medicinesResult, protocolsResult, logsResult, stockResult] = await Promise.all([
-    supabase
-      .from('medicines')
-      .select('id, name, dosage_per_pill, dosage_unit, active_ingredient, therapeutic_class')
-      .eq('user_id', userId),
+  const [medicinesResult, protocolsResult, logsResult, stockResult] = await _fetchRawPatientData(userId, yesterday)
 
-    supabase
-      .from('protocols')
-      .select('id, medicine_id, frequency, time_schedule, dosage_per_intake')
-      .eq('user_id', userId)
-      .eq('active', true),
-
-    supabase
-      .from('medicine_logs')
-      .select('protocol_id, taken_at')
-      .eq('user_id', userId)
-      .gte('taken_at', yesterday),
-
-    supabase
-      .from('stock')
-      .select('medicine_id, quantity')
-      .eq('user_id', userId)
-      .gt('quantity', 0),
-  ])
-
-  logger.debug('✅ Queries Supabase completadas', {
-    userId,
-    medicinesCount: medicinesResult.data?.length || 0,
-    medicinesError: medicinesResult.error?.message,
-    protocolsCount: protocolsResult.data?.length || 0,
-    protocolsError: protocolsResult.error?.message,
-    logsCount: logsResult.data?.length || 0,
-    logsError: logsResult.error?.message,
-    stockCount: stockResult.data?.length || 0,
-    stockError: stockResult.error?.message,
-  })
-
-  if (medicinesResult.error) {
-    const { code, details, hint } = medicinesResult.error
-    logger.error('❌ Erro ao buscar medicamentos', medicinesResult.error, {
-      userId,
-      errorCode: code,
-      errorDetails: details,
-      errorHint: hint,
-    })
-    throw medicinesResult.error
-  }
+  _logQueryResults(userId, medicinesResult, protocolsResult, logsResult, stockResult)
+  _checkMedicineError(userId, medicinesResult)
 
   // Filtrar logs de hoje (com timezone correto)
-  const todayLogs = (logsResult.data || []).filter(log => {
-    return getTodayLocal(getSaoPauloTime(log.taken_at)) === todayStr
-  })
+  const todayLogs = _filterTodayLogs(logsResult.data, todayStr)
 
   // Agregar estoque por medicamento
-  const stockByMedicine = {}
-  for (const entry of (stockResult.data || [])) {
-    stockByMedicine[entry.medicine_id] = (stockByMedicine[entry.medicine_id] || 0) + entry.quantity
-  }
-
-  const stockSummary = Object.entries(stockByMedicine).map(([medicine_id, quantity]) => ({
-    medicine_id,
-    quantity,
-  }))
+  const stockSummary = _aggregateStock(stockResult.data)
 
   // Calcular adesão simples dos últimos 7 dias
   const stats = await calculateSimpleAdherence(userId, protocolsResult.data || [])
@@ -354,6 +371,102 @@ export function updateConversationHistory(userId, userMessage, assistantResponse
 
 // -- Função principal --
 
+async function _fetchAndLogPatientData(userId) {
+  try {
+    logger.info('📊 Buscando dados do paciente no Supabase', { userId })
+    const patientData = await fetchPatientData(userId)
+    logger.info('✅ Dados do paciente obtidos com sucesso', {
+      userId,
+      medicinesCount: patientData.medicines?.length || 0,
+      protocolsCount: patientData.protocols?.length || 0,
+      logsCount: patientData.logs?.length || 0,
+      stockCount: patientData.stockSummary?.length || 0,
+      adherencePercent: patientData.stats?.adherence
+        ? Math.round(patientData.stats.adherence * 100)
+        : null,
+    })
+    return { patientData, error: null }
+  } catch (error) {
+    logger.error('❌ Erro ao buscar dados do paciente', error, { userId })
+    return {
+      patientData: null,
+      error: 'Desculpe, tive um problema ao carregar seus dados. Tente novamente.',
+    }
+  }
+}
+
+async function _callGroqApi(userId, messages) {
+  try {
+    const groq = new Groq({ apiKey: process.env.GROQ_API_KEY })
+
+    logger.info('🔵 Chamando Groq API', {
+      userId,
+      model: MODEL,
+      messagesCount: messages.length,
+      maxTokens: CHATBOT_MAX_TOKENS,
+      temperature: CHATBOT_TEMPERATURE,
+      topP: CHATBOT_TOP_P,
+    })
+
+    const completion = await groq.chat.completions.create({
+      model: MODEL,
+      messages,
+      max_tokens: CHATBOT_MAX_TOKENS,
+      temperature: CHATBOT_TEMPERATURE,
+      top_p: CHATBOT_TOP_P,
+    })
+    return { completion, error: null }
+  } catch (error) {
+    logger.error('❌ Erro ao chamar Groq API', error, {
+      userId,
+      errorStatus: error.status,
+      errorMessage: error.message,
+      errorType: error.constructor.name,
+    })
+
+    if (error.status === 429) {
+      logger.warn('⚠️ Rate limit na Groq API', { userId })
+      return {
+        completion: null,
+        error: '⏱️ Serviço de IA sobrecarregado. Tente novamente em alguns segundos.',
+      }
+    }
+
+    return {
+      completion: null,
+      error: 'Desculpe, estou com dificuldades técnicas. Tente novamente em instantes.',
+    }
+  }
+}
+
+function _checkPreconditions(message, userId) {
+  const validation = validateServerMessage(message)
+  if (validation.blocked) {
+    logger.info('⛔ Mensagem bloqueada pela validação', { userId, reason: validation.reason })
+    return { error: validation.reason, blocked: true, rateLimited: false }
+  }
+
+  if (isServerRateLimited(userId)) {
+    logger.warn('⏱️ Rate limit atingido para usuário', { userId })
+    return {
+      error: '⏱️ Muitas perguntas! Aguarde alguns minutos e tente novamente.',
+      blocked: false,
+      rateLimited: true,
+    }
+  }
+
+  if (!process.env.GROQ_API_KEY) {
+    logger.error('❌ GROQ_API_KEY não configurada')
+    return {
+      error: '🤖 Assistente IA temporariamente indisponível.',
+      blocked: false,
+      rateLimited: false,
+    }
+  }
+
+  return { error: null }
+}
+
 /**
  * Envia mensagem ao chatbot IA via Groq e retorna resposta.
  *
@@ -374,61 +487,15 @@ export async function sendTelegramChatMessage({ message, userId }) {
     msgPreview: message.substring(0, 50),
   })
 
-  // 1. Validar mensagem
-  logger.debug('🔍 Validando mensagem', { userId })
-  const validation = validateServerMessage(message)
-  if (validation.blocked) {
-    logger.info('⛔ Mensagem bloqueada pela validação', {
-      userId,
-      reason: validation.reason,
-    })
-    return { response: validation.reason, blocked: true, rateLimited: false }
-  }
-  logger.debug('✅ Validação passou', { userId })
-
-  // 2. Rate limiting
-  if (isServerRateLimited(userId)) {
-    logger.warn('⏱️ Rate limit atingido para usuário', { userId })
-    return {
-      response: '⏱️ Muitas perguntas! Aguarde alguns minutos e tente novamente.',
-      blocked: false,
-      rateLimited: true,
-    }
-  }
-  logger.debug('✅ Rate limit OK', { userId })
-
-  // 3. Verificar API key
-  if (!process.env.GROQ_API_KEY) {
-    logger.error('❌ GROQ_API_KEY não configurada')
-    return {
-      response: '🤖 Assistente IA temporariamente indisponível.',
-      blocked: false,
-      rateLimited: false,
-    }
+  const { error: preconditionError, blocked, rateLimited } = _checkPreconditions(message, userId)
+  if (preconditionError) {
+    return { response: preconditionError, blocked, rateLimited }
   }
 
   // 4. Buscar dados do paciente
-  let patientData
-  try {
-    logger.info('📊 Buscando dados do paciente no Supabase', { userId })
-    patientData = await fetchPatientData(userId)
-    logger.info('✅ Dados do paciente obtidos com sucesso', {
-      userId,
-      medicinesCount: patientData.medicines?.length || 0,
-      protocolsCount: patientData.protocols?.length || 0,
-      logsCount: patientData.logs?.length || 0,
-      stockCount: patientData.stockSummary?.length || 0,
-      adherencePercent: patientData.stats?.adherence
-        ? Math.round(patientData.stats.adherence * 100)
-        : null,
-    })
-  } catch (error) {
-    logger.error('❌ Erro ao buscar dados do paciente', error, { userId })
-    return {
-      response: 'Desculpe, tive um problema ao carregar seus dados. Tente novamente.',
-      blocked: false,
-      rateLimited: false,
-    }
+  const { patientData, error: fetchError } = await _fetchAndLogPatientData(userId)
+  if (fetchError) {
+    return { response: fetchError, blocked: false, rateLimited: false }
   }
 
   // 5. Construir contexto e system prompt
@@ -444,90 +511,52 @@ export async function sendTelegramChatMessage({ message, userId }) {
   })
 
   // 6. Chamar Groq API
-  try {
-    const groq = new Groq({ apiKey: process.env.GROQ_API_KEY })
-    const messages = [
-      { role: 'system', content: systemPrompt },
-      ...history,
-      { role: 'user', content: message },
-    ]
-
-    logger.info('🔵 Chamando Groq API', {
-      userId,
-      model: MODEL,
-      messagesCount: messages.length,
-      maxTokens: CHATBOT_MAX_TOKENS,
-      temperature: CHATBOT_TEMPERATURE,
-      topP: CHATBOT_TOP_P,
-    })
-
-    const completion = await groq.chat.completions.create({
-      model: MODEL,
-      messages,
-      max_tokens: CHATBOT_MAX_TOKENS,
-      temperature: CHATBOT_TEMPERATURE,
-      top_p: CHATBOT_TOP_P,
-    })
-
-    const rawResponse =
-      completion.choices[0]?.message?.content || 'Desculpe, não consegui responder.'
-
-    const promptTokens = completion.usage?.prompt_tokens || 0
-    const cachedTokens = completion.usage?.cached_prompt_tokens || 0
-    const cacheHitRate = promptTokens > 0 ? Math.round((cachedTokens / promptTokens) * 100) : 0
-    const estimatedSavings = Math.round(cachedTokens * 0.5) // 50% desconto em cached_tokens
-
-    logger.info('✅ Groq respondeu com sucesso', {
-      userId,
-      rawResponseLen: rawResponse.length,
-      promptTokens,
-      cachedTokens,
-      cacheHitRate: `${cacheHitRate}%`,
-      estimatedTokenSavings: estimatedSavings,
-      completionTokens: completion.usage?.completion_tokens,
-      totalTokens: completion.usage?.total_tokens,
-    })
-
-    // 7. Adicionar disclaimer se necessário
-    const response = addServerDisclaimer(rawResponse)
-    logger.debug('📝 Disclaimer adicionado (se necessário)', {
-      userId,
-      finalLen: response.length,
-      disclaimerAdded: response.length > rawResponse.length,
-    })
-
-    // 8. Incrementar rate counter e atualizar histórico
-    incrementServerRateCounter(userId)
-    updateConversationHistory(userId, message, response)
-    logger.info('✅ Taxa incrementada e histórico atualizado', { userId })
-
-    logger.info('✅ Resposta final pronta para envio', {
-      userId,
-      finalResponseLen: response.length,
-    })
-
-    return { response, blocked: false, rateLimited: false }
-  } catch (error) {
-    logger.error('❌ Erro ao chamar Groq API', error, {
-      userId,
-      errorStatus: error.status,
-      errorMessage: error.message,
-      errorType: error.constructor.name,
-    })
-
-    if (error.status === 429) {
-      logger.warn('⚠️ Rate limit na Groq API', { userId })
-      return {
-        response: '⏱️ Serviço de IA sobrecarregado. Tente novamente em alguns segundos.',
-        blocked: false,
-        rateLimited: false,
-      }
-    }
-
-    return {
-      response: 'Desculpe, estou com dificuldades técnicas. Tente novamente em instantes.',
-      blocked: false,
-      rateLimited: false,
-    }
+  const messages = [
+    { role: 'system', content: systemPrompt },
+    ...history,
+    { role: 'user', content: message },
+  ]
+  const { completion, error: apiError } = await _callGroqApi(userId, messages)
+  if (apiError) {
+    return { response: apiError, blocked: false, rateLimited: false }
   }
+
+  const rawResponse =
+    completion.choices[0]?.message?.content || 'Desculpe, não consegui responder.'
+
+  const promptTokens = completion.usage?.prompt_tokens || 0
+  const cachedTokens = completion.usage?.cached_prompt_tokens || 0
+  const cacheHitRate = promptTokens > 0 ? Math.round((cachedTokens / promptTokens) * 100) : 0
+  const estimatedSavings = Math.round(cachedTokens * 0.5) // 50% desconto em cached_tokens
+
+  logger.info('✅ Groq respondeu com sucesso', {
+    userId,
+    rawResponseLen: rawResponse.length,
+    promptTokens,
+    cachedTokens,
+    cacheHitRate: `${cacheHitRate}%`,
+    estimatedTokenSavings: estimatedSavings,
+    completionTokens: completion.usage?.completion_tokens,
+    totalTokens: completion.usage?.total_tokens,
+  })
+
+  // 7. Adicionar disclaimer se necessário
+  const response = addServerDisclaimer(rawResponse)
+  logger.debug('📝 Disclaimer adicionado (se necessário)', {
+    userId,
+    finalLen: response.length,
+    disclaimerAdded: response.length > rawResponse.length,
+  })
+
+  // 8. Incrementar rate counter e atualizar histórico
+  incrementServerRateCounter(userId)
+  updateConversationHistory(userId, message, response)
+  logger.info('✅ Taxa incrementada e histórico atualizado', { userId })
+
+  logger.info('✅ Resposta final pronta para envio', {
+    userId,
+    finalResponseLen: response.length,
+  })
+
+  return { response, blocked: false, rateLimited: false }
 }
